add optional `-s sender' option to beep client
This enables beepagePHP to pass the contents of the
REMOTE_USER server variable to the beep client running on
localhost. Consequently, the cell phone or pager that receives
the page displays the name of the originating sender and not the
name of the httpd user account from the web server.

I had originally uploaded this code on 2004-Jul-28
to http://sourceforge.net/projects/beepagephp/